Simplify donation verification process
Don't do anything intensive on the client, just send the Discord ID (if applicable) through to PayPal and wait for it on the server's end to verify who sent it. This reduces the logic required and is just as effective, if not a bit safer. The client will still have to manually refresh the donation list afterward which may prove problematic since IPNs are not necessarily 'instant'.